+++ NApi/Types/JsString.cs
@@ -19,13 +19,13 @@ namespace NApi.Types
 
     public override unsafe string ToString()
     {
-      ulong length;
+      nuint length;
       NApi.ApiProvider.JsNativeApi.napi_get_value_string_utf8(Scope.Env.EnvPtr, ValuePtr, null, 0, &length).ThrowIfFailed(Scope);
 
       sbyte[] buf = new sbyte[length + 1];
       fixed (sbyte* bufStart = &buf[0])
       {
-        NApi.ApiProvider.JsNativeApi.napi_get_value_string_utf8(Scope.Env.EnvPtr, ValuePtr, bufStart, (ulong)buf.Length, null).ThrowIfFailed(Scope);
+        NApi.ApiProvider.JsNativeApi.napi_get_value_string_utf8(Scope.Env.EnvPtr, ValuePtr, bufStart, (nuint)buf.Length, null).ThrowIfFailed(Scope);
         return new string(bufStart, 0, buf.Length - 1, Encoding.UTF8);
       }
     }
